feat(apollo-vertex): component refactor — InsightCardInner, chart stubs, data context
